refactor: import webhook object from manager
